pyproject.toml
[tool.poetry]
name = "websitechecker"
version = "0.1.0"
description = ""
authors = ["Chiara Mezzavilla <me@chiaramezzavilla.it>"]
readme = "README.md"
packages = [{include = "websitechecker"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.1"
kafka-python = "^2.0.2"
psycopg2-binary = "^2.9.5"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.0"
isort = "^5.11.4"
flake8 = "^6.0.0"
mypy = "^0.991"
pre-commit = "^2.21.0"
responses = "^0.22.0"
types-requests = "^2.28.11.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
